bugff

+++ app/src/main/java/com/madaex/exchange/ui/buy/bean/DealBean2.java
@@ -1,6 +1,5 @@
 package com.madaex.exchange.ui.buy.bean;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -23,8 +22,8 @@ public class DealBean2 {
     private String channel;
     private String dataType;
     private int timestamp;
-    private List<List<BigDecimal>> asks;
-    private List<List<BigDecimal>> bids;
+    private List<List<Float>> asks;
+    private List<List<Float>> bids;
 
     public String getChannel() {
         return channel;
@@ -50,19 +49,19 @@ public class DealBean2 {
         this.timestamp = timestamp;
     }
 
-    public List<List<BigDecimal>> getAsks() {
+    public List<List<Float>> getAsks() {
         return asks;
     }
 
-    public void setAsks(List<List<BigDecimal>> asks) {
+    public void setAsks(List<List<Float>> asks) {
         this.asks = asks;
     }
 
-    public List<List<BigDecimal>> getBids() {
+    public List<List<Float>> getBids() {
         return bids;
     }
 
-    public void setBids(List<List<BigDecimal>> bids) {
+    public void setBids(List<List<Float>> bids) {
         this.bids = bids;
     }
 }
